continue working on yaml command file
