+++ go.mod
@@ -5,9 +5,12 @@ go 1.17
 require (
 	github.com/golang/glog v1.0.0
 	golang.org/x/crypto v0.0.0-20220314234724-5d542ad81a58
+	gopkg.in/yaml.v2 v2.4.0
 )
 
 require (
-	github.com/sbezverk/gobmp v1.0.1 // indirect
+	github.com/kr/text v0.2.0 // indirect
+	github.com/niemeyer/pretty v0.0.0-20200227124842-a10e7caefd8e // indirect
 	golang.org/x/sys v0.0.0-20210615035016-665e8c7367d1 // indirect
+	gopkg.in/check.v1 v1.0.0-20200227125254-8fa46927fb4f // indirect
 )
