+++ pkg/patterns/patterns_test.go
